refactor: improve abstract integration class
